refactor: streamline component architecture & remove unused code
