+++ src/components/display/index.ts
@@ -5,5 +5,4 @@ export { VersionBadge } from './VersionBadge';
 export { PaginationDots } from './PaginationDots';
 export { SectionHeading } from './SectionHeading';
 export { SkillPill } from './SkillPill';
-export { ScrollIndicator } from './ScrollIndicator';
 export { default as FloatingParticle } from './FloatingParticle';
