+++ src/components/interactive/index.ts
@@ -1,2 +1,1 @@
-export { ThemeButton } from './ThemeButton';
 export { SocialIcons } from './SocialIcons';
