--- src/data/content/contact.tsx
@@ -1,7 +1,0 @@
-// src/data/content/contact.tsx
-
-// contact content
-export const contactContent = {
-    intro: `I am currently open to new opportunities. Whether you have a question, a project idea, 
-            or just want to say hello, feel free to reach out!`
-  };
