+++ src/sections/shared/ResponsiveWrapper.tsx
@@ -10,11 +10,6 @@ export interface ResponsiveWrapperProps {
   className?: string;
 }
 
-export interface ConditionalRenderProps {
-  children: ReactNode;
-  showOn?: 'mobile' | 'tablet' | 'desktop' | 'mobile-tablet' | 'tablet-desktop' | 'not-mobile' | 'not-desktop';
-  className?: string;
-}
 
 // Main responsive wrapper that renders different content based on screen size
 export function ResponsiveWrapper({ 
@@ -38,39 +33,6 @@ export function ResponsiveWrapper({
   return <div className={wrapperClasses}>{desktop}</div>;
 }
 
-// Conditional renderer for specific screen sizes
-export function ConditionalRender({ 
-  children, 
-  showOn = 'desktop', 
-  className = '' 
-}: ConditionalRenderProps) {
-  const { shouldShowMobileLayout, shouldShowTabletLayout, shouldShowDesktopLayout } = useResponsiveSection();
-
-  const shouldRender = (() => {
-    switch (showOn) {
-      case 'mobile':
-        return shouldShowMobileLayout;
-      case 'tablet':
-        return shouldShowTabletLayout;
-      case 'desktop':
-        return shouldShowDesktopLayout;
-      case 'mobile-tablet':
-        return shouldShowMobileLayout || shouldShowTabletLayout;
-      case 'tablet-desktop':
-        return shouldShowTabletLayout || shouldShowDesktopLayout;
-      case 'not-mobile':
-        return !shouldShowMobileLayout;
-      case 'not-desktop':
-        return !shouldShowDesktopLayout;
-      default:
-        return true;
-    }
-  })();
-
-  if (!shouldRender) return null;
-
-  return className ? <div className={className}>{children}</div> : <>{children}</>;
-}
 
 // Specialized wrappers for common patterns
 export interface TimelineResponsiveProps {
@@ -115,26 +77,3 @@ export function TableResponsive({
   );
 }
 
-export interface NavigationResponsiveProps {
-  sidebarNavigation?: ReactNode;
-  mobileNavigation: ReactNode;
-  className?: string;
-}
-
-export function NavigationResponsive({ 
-  sidebarNavigation, 
-  mobileNavigation, 
-  className = '' 
-}: NavigationResponsiveProps) {
-  const { showSidebarNavigation, showMobileNavigation } = useResponsiveSection();
-
-  if (showMobileNavigation) {
-    return <div className={className}>{mobileNavigation}</div>;
-  }
-  
-  if (showSidebarNavigation && sidebarNavigation) {
-    return <div className={className}>{sidebarNavigation}</div>;
-  }
-  
-  return <div className={className}>{mobileNavigation}</div>;
-}
