+++ src/sections/shared/SectionContainer.tsx
@@ -57,42 +57,3 @@ export function SectionContainer({
   );
 }
 
-// Specialized containers for common section types
-export function HeroSectionContainer({ children, className = '' }: Pick<SectionContainerProps, 'children' | 'className'>) {
-  return (
-    <SectionContainer
-      maxWidth="full"
-      spacing="lg"
-      className={`min-h-screen flex items-center ${className}`}
-      as="section"
-    >
-      {children}
-    </SectionContainer>
-  );
-}
-
-export function ContentSectionContainer({ children, className = '' }: Pick<SectionContainerProps, 'children' | 'className'>) {
-  return (
-    <SectionContainer
-      maxWidth="xl"
-      spacing="md"
-      className={`py-16 ${className}`}
-      as="section"
-    >
-      {children}
-    </SectionContainer>
-  );
-}
-
-export function CompactSectionContainer({ children, className = '' }: Pick<SectionContainerProps, 'children' | 'className'>) {
-  return (
-    <SectionContainer
-      maxWidth="lg"
-      spacing="sm"
-      className={`py-8 ${className}`}
-      as="section"
-    >
-      {children}
-    </SectionContainer>
-  );
-}
